data_parser/requirements.txt
pdf2image>=1.16.0
Pillow>=9.0.0
pytesseract>=0.3.10
# Optional (Python 3.9–3.13 only): grid-based extraction via img2table
# pip install img2table opencv-python-headless  then run pdf_table_to_csv_img2table.py
